pyproject.toml
[project]
name = "givenpy"
version = "1.0.5"
description = "A simple micro BDD framework for Python"
authors = [
    { name = "Tadas Subonis", email = "tadas.subonis@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = []
keywords = ["BDD", "testing", "micro", "framework", "test", "unittest"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
]

[project.urls]
Documentation = "https://github.com/tadas-subonis/givenpy"
Repository = "https://github.com/tadas-subonis/givenpy"
"Bug Tracker" = "https://github.com/tadas-subonis/givenpy/issues"

[dependency-groups]
dev = [
    "pyhamcrest>=2.0.4",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
